Add files via upload
Variation on the legendre formula prime counting algorithm. Although, inefficient when compared with the other implementation, it illustrates finely some of the techniques used in sieve theory like the use of weighting functions to estimatethe size of sifted sets, and the general structure of prime sieve methods in number theory. It can be adapted by using more approppriate weighting functions