Add files via upload
Routines to test the primality of integers using Chebyshev polynomials. It's fast enough to test numbers up to 10^100 in a few seconds. This algorithms give a lot of false negatives, i.e., many primes are declared composite. But no false positive was detected. This algorithm isn't adequate as a primality test per si, but it is very useful to generate industrial primes (combined for example with a Miller-Rabin test or a strong Lucas test). Maybe it needs some improvements in parameter selection.