Add files via upload
Useful routines to approximate the class number of imaginary quadratic fields. The lower bound is very inaccurate, but the upper is quite good differing by C*ln(D) of the correct value, where C is tipically lower than 10. It can be used when it is possible to test the values of h(-D) in a small range and the exact value is not needed