Add files via upload
Refactored code in Python for multiprecision arithmetic and for the use built-in implementation of complimentary error function provided by the Scipy library. Now it can compute class of imaginary quadratic field of fundamental discrminant up to 10^15 in a reasonable time. In tests, about 15% of the time the answer is wrong by a small value, it was checked using PARI/GP and the program using counting of reduced forms.